Created Event Listeners to make articles shrink
There are now event listeners attached to the article elements to make them shrink.

This replaces the css pseudo-selector implementation.

A nice side effect: they now transition smoothly to grow back.

I did this so that clicking the buttons would not make the articles shrink to make it clear that you are pressing a button.