Checks to avoid calling int() on NoneType
calling int() on a NoneType results in a python error.

I renamed the variable that stores the value of "localStorage.getItem(config.localStorage_length_key)" to "num_of_urls_string".

Now the integer value of that string is in "num_of_urls" which is 0 by default and if localStorage.getItem returned a NoneType.