Improved Recommendation (More terms) And Clear Search Terms on Seg Select
The max_terms was too low so I was getting bad recs.

I made it so switching the seg select clears the search terms. The search term messed me up in the demo.

Also the articles are loaded for the recs and the TFID and cosine similarity are calculated at the begining so the site doesn't freeze when clicking for you for the first time.